requirements.txt
Flask
pandas
numpy
setuptools
PyPDF2
sentence-transformers
faiss-cpu
tqdm
pytesseract
pillow
-e .
